+++ index.html
@@ -1235,25 +1235,7 @@
                             </div>
                             <input type="text" name="tone-example-1" class="tone-example"
                                 placeholder="I'm a plumber, not a doctor. Lets-a go!">
-                            <button type="button" class=" material-symbols-outlined btn-add-tone-example">add_circle</button>
-
-
-                        </div>
-                        <div class="step">
-                            <h3>Finalize</h3>
-                            <!-- show a summary -->
-                             <div class="summary">
-                                <div class="card-summary"></div>
-                                <div class="details">
-                                    <p>Prompt: <span class="prompt"></span></p>
-                                    <p>Aggressiveness: <span class="aggressiveness"></span></p>
-                                    <p>Sensuality: <span class="sensuality"></span></p>
-                                    <p>Internet Access: <span class="internet-access"></span></p>
-                                    <p>Roleplay Enabled: <span class="roleplay"></span></p>
-                                </div>
-                             </div>
-
-
+                            <button type="button" id="btn-add-tone-example" class="material-symbols-outlined">add_circle</button>
                         </div>
                     </div>
                     <div class="stepper-footer">
